Metodo de agregar y comprobar lector listo

+++ TallerDePOO/src/main/java/Biblioteca/Lector.java
@@ -7,7 +7,7 @@ public class Lector extends Persona{
 	private Boolean estaMultado;
 	private ArrayList<Ejemplar> ejemplares;
 	
-	public Lector(String nom, String apellido, String tipDoc, int numDoc, String correo, long numCel, LocalDate fecNac,
+	public Lector(String nom, String apellido, String tipDoc, int numDoc, String correo, int numCel, LocalDate fecNac,
 			String sex, String nac, String domi, int codPos, String dep, String localidad, int canMulta,
 			Boolean estaMultado, ArrayList<Ejemplar> ejemplares) {
 		super(nom, apellido, tipDoc, numDoc, correo, numCel, fecNac, sex, nac, domi, codPos, dep, localidad);
@@ -17,7 +17,7 @@ public class Lector extends Persona{
 	}
 	
 	//Constructor para agregar datos del lector
-	public Lector(String nom, String apellido, String tipDoc, int numDoc, String correo, long numCel, LocalDate fecNac,
+	public Lector(String nom, String apellido, String tipDoc, int numDoc, String correo, int numCel, LocalDate fecNac,
 			String sex, String nac, String domi, int codPos, String dep, String localidad) {
 		
 		super(nom, apellido, tipDoc, numDoc, correo, numCel, fecNac, sex, nac, domi, codPos, dep, localidad);
